+++ tagger/src/components/picker_footer.rs
@@ -47,6 +47,7 @@ impl Widget for PickerNavigation {
                     Span::styled("Finished", Style::default().fg(Color::Green))
                 }
                 CurrentScreen::Exiting => Span::styled("Exiting", Style::default().fg(Color::Red)),
+                _ => unreachable!(),
             },
             Span::styled(" | ", Style::default().fg(Color::White)),
             match self.current_screen {
@@ -66,6 +67,7 @@ impl Widget for PickerNavigation {
                 CurrentScreen::Exiting => {
                     Span::styled("Exiting", Style::default().fg(Color::LightRed))
                 }
+                _ => unreachable!(),
             },
         ];
         let mode_footer = Paragraph::new(Line::from(current_navigation_text))
@@ -89,6 +91,7 @@ impl Widget for PickerHint {
                     "Are you sure you want to exit? (y/n)",
                     Style::default().fg(Color::LightRed),
                 ),
+                _ => unreachable!(),
             }
         };
         let key_notes_footer = Paragraph::new(hint).block(Block::default().borders(Borders::ALL));